refactor: flag support added
generate --mean
is now command to generate mean apps.
