+++ lib/utils/flags.js
@@ -1,0 +1,3 @@
+// var args = process.argv.splice(2);
+
+// console.log(args[0])
